Change list: - Install flickity - Switch to carousel for song selection - move controls and score sections - minor refactoring
