+++ index.html
@@ -6,7 +6,7 @@
     <title> V I B E </title>
         <script src="./dist/main.js"></script>
         <link rel="stylesheet" href="./dist/main.css"/>
-        <link href="https://fonts.googleapis.com/css?family=Monoton&display=swap" rel="stylesheet">
+        <link href="https://fonts.googleapis.com/css2?family=Kdam+Thmor+Pro&family=Monoton&family=Roboto:wght@100;300;500&display=swap" rel="stylesheet" />
         <script src="src/three.js"></script>
 </head>
 
@@ -19,21 +19,16 @@
         <div class="controls">
             <ul>
                 <h2>Controls</h2>
-                <li>D for <div class="red-circle"></div></li>
-                <li>F for <div class="green-circle"></div></li>
-                <li>J for <div class="blue-circle"></div></li>
-                <li>K for <div class="yellow-circle"></div></li>
+                <li>D for <div class="red-key circle-keys"></div></li>
+                <li>F for <div class="green-key circle-keys"></div></li>
+                <li>J for <div class="blue-key circle-keys"></div></li>
+                <li>K for <div class="yellow-key circle-keys"></div></li>
                 <li>Space Bar to Pause</li>
             </ul>
         </div>
-        <div class="score">
-            <h3>Score</h3>
-            <h3 id="score-value"></h3>
-        </div>
-
         <div class="song-selection">
             <h3>Please Select A Song</h3>
-            <ul id="song-list"></ul>
+            <div id="song-list" class="main-carousel"></div>
         </div>
 
         <div id="pause-menu" class='hide-menu'>
@@ -42,6 +37,11 @@
                 <li id="restart">Restart</li>
                 <li id="quit">Quit</li>
             </ul>
+        </div>
+
+        <div class="score">
+            <h3>Score</h3>
+            <h3 id="score-value"></h3>
         </div>
         <canvas id='canvas'></canvas>
     </div>
