+++ package.json
@@ -31,6 +31,7 @@
     "autoprefixer": "^9.6.5",
     "babel-loader": "^8.0.6",
     "css-loader": "^3.2.0",
+    "flickity": "^3.0.0",
     "mini-css-extract-plugin": "^1.3.9",
     "postcss-loader": "^3.0.0",
     "sass": "^1.69.5",
