+++ src/js/controls.js
@@ -15,26 +15,26 @@ class Controls {
       font: font,
       size: 1,
       height: 1,
-      curveSegments: 3
+      curvedSegments: 3,
     });
 
     let materials = [
-      new THREE.MeshPhongMaterial({ color: 0x2690f7, flatShading: true }), // front
-      new THREE.MeshPhongMaterial({ color: 0x064d92 }) // side
+      new THREE.MeshPhongMaterial({ color: "yellow" }), // front
+      new THREE.MeshPhongMaterial({ color: 'blue', flatShading: true }), // side
     ];
 
     this.titleRender = new THREE.Mesh(title, materials);
     this.titleRender.name = "title";
 
-    this.titleRender.position.z = -4;
-    this.titleRender.position.x = -2.25;
+    // this.titleRender.position.z = -4;
+    this.titleRender.position.x = -2;
     this.titleRender.position.y = 5;
     // titleRender.rotation.x = 0.5;
     // titleRender.rotation.y = 0.5;
     this.scene.add(this.titleRender);
 
     let light = new THREE.DirectionalLight(0xffffff);
-    light.position.set(2, 3, 2);
+    light.position.set(0, -1, 2);
     window.light = light;
     this.scene.add(light);
   }
