+++ src/js/game.js
@@ -1,5 +1,6 @@
 import Visualizer from './visualizer';
 import AudioSynth from "./audio";
+import Flickity from 'flickity';
 
 class Game {
   constructor() {
@@ -77,15 +78,20 @@ class Game {
     ];
     audioList.forEach((songData) => {
       let [songUrl, number] = songData;
-      let listItem = document.createElement("li");
+      let listItem = document.createElement("div");
       listItem.setAttribute('id', `song-id-${number}`);
-      listItem.setAttribute('class', 'cards');
+      listItem.setAttribute('class', 'carousel-cell');
+      let imgTag = document.createElement('img');
+      imgTag.src = `assets/images/album art/${songUrl}.jpeg`;
+      listItem.appendChild(imgTag);
       listItem.append(songUrl);
       listItem.onclick = () => {
         this.curSongUrl = songUrl;
         this.playRound(songUrl)};
       songList.append(listItem);
     });
+
+    new Flickity(songList, {wrapAround: true, accessibility: false})
   }
 
   setMenuEvents() {
@@ -106,7 +112,6 @@ class Game {
   static scoreUpdate(score) {
     let scoreCard = document.getElementById("score-value");
     scoreCard.innerText = score;
-    // console.log(this.score)
   }
 
   visualInit() {
