+++ src/js/sphere.js
@@ -8,39 +8,30 @@ class Sphere {
     });
     this.sphereShape = new THREE.Mesh(this.geometry, this.material);
     this.sphereShape.dynamic = true;
-      let numSides = this.geometry.faces.length;
-      for (let i = 0; i < numSides; i++) {
-        let vertIndex = this.geometry.faces[i];
-        var color = new THREE.Color(0xffffff);
-        let rRandom = Math.random() * (1 - 0.64) + 0.64;
-        let gRandom = Math.random() * (1 - 0.64) + 0.64;
-        let bRandom = Math.random() * (1 - 0.64) + 0.64;
-        // console.log([rRandom, gRandom, bRandom]);
-        color.setRGB(rRandom, gRandom, bRandom);
-        vertIndex.color = color;
-      }
-      this.geometry.elementsNeedUpdate = true;
+
+    this.generateRandomSphereColors();
+
     this.sphereShape.position.y = 2.5;
   }
 
-  colorSwitch(maxFreq){
+  generateRandomSphereColors(){
+    let numSides = this.geometry.faces.length;
+    for (let i = 0; i < numSides; i++) {
+      let vertIndex = this.geometry.faces[i];
+      var color = new THREE.Color(0xffffff);
+      let rRandom = Math.random() * (1 - 0.64) + 0.64;
+      let gRandom = Math.random() * (1 - 0.64) + 0.64;
+      let bRandom = Math.random() * (1 - 0.64) + 0.64;
+      color.setRGB(rRandom, gRandom, bRandom);
+      vertIndex.color = color;
+    }
+    this.geometry.elementsNeedUpdate = true;
+  }
 
-    
-      if(maxFreq === 255){
-        let numSides = this.geometry.faces.length;
-        for (let i = 0; i < numSides; i++) {
-          let vertIndex = this.geometry.faces[i];
-          var color = new THREE.Color(0xffffff);
-          let rRandom = Math.random() * (1 - 0.64) + 0.64;
-          let gRandom = Math.random() * (1 - 0.64) + 0.64;
-          let bRandom = Math.random() * (1 - 0.64) + 0.64;
-          // console.log([rRandom,gRandom,bRandom])
-          color.setRGB(rRandom, gRandom, bRandom);
-          vertIndex.color = color;
-        }
-        
-      }
-      this.geometry.elementsNeedUpdate = true;
+  colorSwitch(maxFreq){
+    if(maxFreq === 255){
+      this.generateRandomSphereColors 
+    }
   }
 }
