+++ src/shared/variables.js
