+++ src/styles/variables.scss
@@ -1,0 +1,8 @@
+// key colors
+$red: #ff0000;
+$blue: #0000ff;
+$yellow: #ffff00;
+$green: #00ff00;
+
+// others
+$gray: #e3dbdbad;
